requirements.txt
pycryptodome==3.14.1
pyOpenSSL==22.0.0
requests==2.28.0
